backend/requirements.txt
absl-py==2.2.2
amqp==5.3.1
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django_celery_results==2.6.0
djangorestframework==3.16.0
ecdsa==0.19.1
flatbuffers==25.2.10
fonttools==4.57.0
frozendict==2.4.6
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
joblib==1.4.2
keras==3.9.2
kiwisolver==1.4.8
kombu==5.5.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
multitasking==0.0.11
namex==0.0.8
numpy==2.1.3
nvidia-nccl-cu12==2.26.2.post1
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.50
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
rich==14.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xgboost==3.0.0
yfinance==0.2.55
